Add typecast to removing a compiler warning on systems that consider mode_t and %o different.
PR:
Obtained from:
Submitted by:
Reviewed by:	Roy T. Fielding, Jim Jagielski, Dirk-Willem van Gulik


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@79614 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/standard/mod_mime_magic.c
@@ -1530,7 +1530,7 @@ static int fsmagic(request_rec *r, const char *fn, struct stat *sb)
 	break;
     default:
 	aplog_error(APLOG_MARK, APLOG_NOERRNO | APLOG_ERR, r->server,
-		    MODNAME ": invalid mode 0%o.", sb->st_mode);
+		    MODNAME ": invalid mode 0%o.", (unsigned int)sb->st_mode);
 	return HTTP_INTERNAL_SERVER_ERROR;
     }
